34 Добавил 4 функции в одной (для промежуточного задания) Одна функция, 4 разных запроса * получение списка городов по указанному региону * получение списка городов по указанному округу * получения списка городов по указанному диапазону численности населения * получения списка городов по указанному диапазону года основания

+++ 34/pkg/model/city.go
@@ -6,5 +6,5 @@ type City struct {
 	Region     string `csv:"region" json:"region"`
 	District   string `csv:"district" json:"district"`
 	Population uint   `csv:"population" json:"population"`
-	Foundation uint16 `csv:"foundation" json:"foundation"`
+	Foundation uint   `csv:"foundation" json:"foundation"`
 }
